First steps for automatically generating navmesh
This change generates the convex polygons of a navmesh from the collison
geometry. It seems to generate these polygons correctly, but I haven't
seriously verified yet. I still need to somehow store the connections
between polygons, and for this I think I will need to use a different
data structure.